fix: break cron run log parser type cycle
