+++ src/cron/run-log-types.ts
@@ -1,0 +1,31 @@
+import type { FailoverReason } from "../agents/embedded-agent-helpers/types.js";
+import type {
+  CronDeliveryStatus,
+  CronDeliveryTrace,
+  CronFailureNotificationDelivery,
+  CronRunDiagnostics,
+  CronRunStatus,
+  CronRunTelemetry,
+} from "./types.js";
+
+export type CronRunLogEntry = {
+  ts: number;
+  jobId: string;
+  action: "finished";
+  status?: CronRunStatus;
+  error?: string;
+  errorReason?: FailoverReason;
+  summary?: string;
+  diagnostics?: CronRunDiagnostics;
+  delivered?: boolean;
+  deliveryStatus?: CronDeliveryStatus;
+  deliveryError?: string;
+  failureNotificationDelivery?: CronFailureNotificationDelivery;
+  delivery?: CronDeliveryTrace;
+  sessionId?: string;
+  sessionKey?: string;
+  runId?: string;
+  runAtMs?: number;
+  durationMs?: number;
+  nextRunAtMs?: number;
+} & CronRunTelemetry;
